+++ Modules/Intent.Modules.OpenTelemetry/release-notes.md
@@ -1,3 +1,7 @@
+### Version 2.2.11
+
+- Improvement: Updated NuGet package versions.
+
 ### Version 2.2.10
 
 - Improvement: Internal changes around better supporting multiple application entry points.